pyproject.toml
[project]
name = "backaryAPI"
version = "0.1.0"
description = "API for shop"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
